+++ src/providers/ThumbnailsProvider.js
@@ -1,13 +1,24 @@
 "use strict";
 
-function test() {
-  chrome.runtime.sendMessage({ getThumbs: "" }, (thumbs) => {
-    this.setState({
-      thumbs: thumbs
-    })
+import HistoryProvider from './HistoryProvider';
+import StorageProvider from './StorageProvider';
+
+const listeners = []
+
+function init() {
+  chrome.runtime.sendMessage({ getThumbs: "simple" }, (thumbs) => {
+
   });
 }
 
+function onChange(callback) {
+  listeners.push(callback);
+}
+
+function emit(data) {
+  listeners.forEach(listener => { listener(data) })
+}
+
 module.exports = {
-  test: test
+  onChange: onChange
 }